Logic navigation 🚀

+++ Frontend/src/app/components/admin/admin.component.ts
@@ -1,5 +1,7 @@
 import { Component, Injectable, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
 import { EventService } from 'src/app/services/event.service';
+import { UserService } from 'src/app/services/user.service';
 import { Events } from '../../models/event.model';
 
 
@@ -13,14 +15,26 @@ import { Events } from '../../models/event.model';
 export class AdminComponent implements OnInit {
 
   events: Events[] = [];
-  constructor(private eService: EventService) { } 
+  constructor(private eService: EventService, private uService: UserService, private router: Router) { } 
 
   ngOnInit(){
     this.eService.getEvents(0).subscribe(events => this.events = events)
   }
   deleteEvent(id:number){
-    this.eService.deleteEvent(id).subscribe(() => console.log("user deleted"));
+    if (this.uService.isAdmin()){
+      this.eService.deleteEvent(id).subscribe(() => console.log("user deleted"));
+    }else{
+      this.router.navigate([""]);
+    }
+    
   }
 
+  redirectGraphs(id: number){
+    this.router.navigate(['graph-event/'+id]);
+  }
+
+  redirectEditEvent(id: number){
+    this.router.navigate(["edit-event/"+id]);
+  }
 
 }
